Sanitized/normalized the gravity calc. mess
+ Sim. NumT back to double
+ A test runner script to pre-select --g-mode=R

+++ src/Model/Cfg.hpp
@@ -3,7 +3,7 @@
 
 namespace Model {
 
-	using BasicNumberType = float;
+	using BasicNumberType = double;
 
 }
